+++ src/api/controllers/interactions.ts
@@ -1,9 +1,7 @@
+import { DiscordPayload } from "../../types/types.ts";
 import { eventHandlers } from "../../bot.ts";
-import {
-  DiscordPayload,
-  InteractionCommandPayload,
-} from "../../types/types.ts";
-import { createMember } from "../structures/member.ts";
+import { structures } from "../structures/mod.ts";
+import { InteractionCommandPayload } from "../../types/types.ts";
 
 export async function handleInternalInteractionsCreate(data: DiscordPayload) {
   if (data.t !== "INTERACTION_CREATE") return;
@@ -13,7 +11,7 @@ export async function handleInternalInteractionsCreate(data: DiscordPayload) {
   eventHandlers.interactionCreate?.(
     {
       ...payload,
-      member: await createMember(payload.member, payload.guild_id),
+      member: await structures.createMember(payload.member, payload.guild_id),
     },
   );
 }
