+++ src/api/controllers/roles.ts
@@ -1,10 +1,10 @@
 import { eventHandlers } from "../../bot.ts";
+import { structures } from "../structures/structures.ts";
 import {
   DiscordPayload,
   GuildRoleDeletePayload,
   GuildRolePayload,
 } from "../../types/types.ts";
-import { createRole } from "../structures/role.ts";
 import { cacheHandlers } from "./cache.ts";
 
 export async function handleInternalGuildRoleCreate(data: DiscordPayload) {
@@ -14,7 +14,7 @@ export async function handleInternalGuildRoleCreate(data: DiscordPayload) {
   const guild = await cacheHandlers.get("guilds", payload.guild_id);
   if (!guild) return;
 
-  const role = await createRole(payload.role);
+  const role = await structures.createRole(payload.role);
   const roles = guild.roles.set(payload.role.id, role);
   guild.roles = roles;
   return eventHandlers.roleCreate?.(guild, role);
@@ -55,7 +55,7 @@ export async function handleInternalGuildRoleUpdate(data: DiscordPayload) {
   const cachedRole = guild.roles.get(payload.role.id);
   if (!cachedRole) return;
 
-  const role = await createRole(payload.role);
+  const role = await structures.createRole(payload.role);
   guild.roles.set(payload.role.id, role);
   eventHandlers.roleUpdate?.(guild, role, cachedRole);
 }
