+++ src/api/structures/structures.ts
@@ -1,0 +1,7 @@
+export * from "./channel.ts";
+export * from "./guild.ts";
+export * from "./member.ts";
+export * from "./message.ts";
+export * from "./mod.ts";
+export * from "./role.ts";
+export * from "./template.ts";
